Diplom 2017
My diploma is devoted to the topic of viruses. Viruses are the basis of life, even its essence. It is interesting to note that almost all evolutionary changes in living organisms occur as a result of the influence of viruses: vertical and horizontal transfer of genes - their "business". They have created existing life forms, they are transforming it, and they are creating it.